refactor: create judgeme utility functions and update imports for reviews in components for better organization and type clarity
