+++ app/routes/($locale).products.$handle.tsx
@@ -19,7 +19,7 @@ import {
   RECOMMENDED_PRODUCTS_QUERY,
   VARIANTS_QUERY,
 } from '~/data/queries';
-import {createJudgemeReview, getJudgemeReviews} from '~/lib/judgeme';
+import {createJudgemeReview, getJudgemeReviews} from '~/lib/utils/judgeme';
 import {seoPayload} from '~/lib/seo.server';
 import type {Storefront} from '~/lib/types/type-locale';
 import {WeaverseContent} from '~/weaverse';
